# Backend/src/openapi.yaml
openapi: 3.1.0
info:
  title: Kaleem AI - Analytics API
  version: 1.0.0
  description: >
    مواصفة OpenAPI 3.1 لوحدات التحليلات (Analytics) كما في الكود المرفق.
    تعتمد هذه المواصفة فقط على DTOs والواجهات الموجودة. أي حقول غير معرّفة في الشيفرة وُسمت بـ TODO.
servers:
  - url: http://localhost:3000/api
    description: Local
  - url: https://api.kaleem-ai.com/api
    description: Production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PeriodParam:
      name: period
      in: query
      required: false
      schema:
        type: string
        enum: [week, month, quarter]
        default: week
      example: week
      description: الفترة الزمنية للتجميع.
    GroupByParam:
      name: groupBy
      in: query
      required: false
      schema:
        type: string
        enum: [day, hour]
        default: day
      example: day
      description: أسلوب التجميع الزمني.
    PageParam:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
      required: false
      description: رقم الصفحة (1+).
    LimitParam:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, default: 20 }
      required: false
      description: حجم الصفحة.
    LimitTopKeywords:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 10 }
      example: 10
...
    ResolvedParam:
      name: resolved
      in: query
      required: false
      schema:
        type: string
        enum: [all, 'true', 'false']
        default: all
...
  headers:
    X-RateLimit-Limit:
      description: الحد الأقصى للطلبات في النافذة الحالية.
      schema: { type: integer }
    X-RateLimit-Remaining:
      description: عدد الطلبات المتبقية في النافذة الحالية.
      schema: { type: integer }
    X-RateLimit-Reset:
      description: وقت إعادة الضبط بالثواني منذ Unix epoch.
      schema: { type: integer }
  schemas:
...
    KeywordCount:
      type: object
      properties:
        keyword: { type: string, example: 'delivery' }
        count: { type: integer, example: 42 }
      required: [keyword, count]
...
    Overview:
      type: object
      properties:
        sessions:
          type: object
          properties:
            count: { type: integer, example: 560 }
            changePercent:
              type: integer
              nullable: true
              example: 12
          required: [count]
        messages: { type: integer, example: 1345 }
        topKeywords:
          type: array
          items: { $ref: '#/components/schemas/KeywordCount' }
        topProducts:
          type: array
          items: { $ref: '#/components/schemas/TopProduct' }
        channels:
          type: object
          properties:
            total: { type: integer, example: 3 }
            breakdown:
              type: array
              items: { $ref: '#/components/schemas/ChannelCount' }
          required: [total, breakdown]
        orders:
          type: object
          properties:
            count: { type: integer, example: 190 }
            changePercent: { type: integer, nullable: true, example: -5 }
            byStatus:
              type: object
              additionalProperties: { type: integer }
              example:
                created: 50
                paid: 120
                canceled: 20
            totalSales: { type: number, example: 24500.75 }
          required: [count, byStatus, totalSales]
        csat: { type: number, nullable: true, example: 4.6 }
        firstResponseTimeSec: { type: number, nullable: true, example: 42 }
        missingOpen: { type: integer, example: 7 }
        storeExtras:
          type: object
          properties:
            paidOrders: { type: integer, example: 120 }
            aov: { type: number, nullable: true, example: 204.17 }
      required: [sessions, messages, topKeywords, topProducts, channels, orders]
...
    CreateMissingResponseDto:
      type: object
      properties:
        merchant:
          type: string
          description: معرف التاجر
          example: merchant_123
        channel:
          type: string
          enum: [telegram, whatsapp, webchat]
          description: القناة التي وردت منها الرسالة
          example: whatsapp
        question:
          type: string
          description: سؤال العميل
          example: 'هل يتوفر لديكم هذا المنتج؟'
        botReply:
          type: string
          description: رد البوت على سؤال العميل
          example: 'عفواً، لم أفهم السؤال.'
        sessionId:
          type: string
          nullable: true
          description: معرف الجلسة (اختياري)
          example: session_xyz
        aiAnalysis:
          type: string
          nullable: true
          description: تحليل الذكاء الاصطناعي للرسالة (اختياري)
          example: 'العميل يستفسر عن توفر منتج.'
        customerId:
          type: string
          nullable: true
          description: معرف العميل (اختياري)
          example: customer_456
        type:
          type: string
          enum: [missing_response, unavailable_product]
          description: نوع الاستجابة المسجلة
          example: missing_response
        resolved:
          type: boolean
          nullable: true
          description: ما إذا كانت الاستجابة قد تمت معالجتها أم لا (اختياري)
          example: false
      required: [merchant, channel, question, botReply, type]
...
    AckWithId:
      type: object
      properties:
        success: { type: boolean, example: true }
        id: { type: string, example: '64f2c1a9e1...' }
      required: [success, id]
...
    MessagesTimelineResponse:
      type: array
      description: 'الهيكل يعتمد على تجميع المستودع. TODO: تعريف الحقول عند التوفر.'
      items:
        type: object
        example:
          bucket: '2025-09-11'
          count: 123
...
security:
  - bearerAuth: []
paths:
  /analytics/overview:
    get:
      tags: [analytics]
      operationId: analytics_overview
      summary: analytics.operations.dashboard.summary
      description: analytics.operations.dashboard.description
      parameters:
        - $ref: '#/components/parameters/PeriodParam'
      responses:
        '200':
          description: OK
          headers:
            X-RateLimit-Limit:
              { $ref: '#/components/headers/X-RateLimit-Limit' }
            X-RateLimit-Remaining:
              { $ref: '#/components/headers/X-RateLimit-Remaining' }
            X-RateLimit-Reset:
              { $ref: '#/components/headers/X-RateLimit-Reset' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Overview' }
              examples:
                example:
                  value:
                    sessions: { count: 560, changePercent: 12 }
                    messages: 1345
                    topKeywords:
                      - { keyword: 'delivery', count: 42 }
                      - { keyword: 'refund', count: 31 }
                    topProducts:
                      - {
                          productId: '64f2...',
                          name: 'Premium Hoodie',
                          count: 17,
                        }
                    channels:
                      total: 3
                      breakdown:
                        - { channel: 'whatsapp', count: 980 }
                        - { channel: 'telegram', count: 230 }
                        - { channel: 'webchat', count: 135 }
                    orders:
                      count: 190
                      changePercent: -5
                      byStatus: { created: 50, paid: 120, canceled: 20 }
                      totalSales: 24500.75
                    csat: 4.6
                    firstResponseTimeSec: 42
                    missingOpen: 7
                    storeExtras: { paidOrders: 120, aov: 204.17 }
      x-slos:
        response_p95_ms: 500
        availability: 99.9
      x-grafana:
        dashboard: 'TODO: https://grafana.example.com/d/analytics-overview'
  /analytics/top-keywords:
    get:
      tags: [analytics]
      operationId: analytics_top_keywords
      summary: analytics.operations.metrics.summary
      description: analytics.operations.metrics.description
      parameters:
        - $ref: '#/components/parameters/PeriodParam'
        - $ref: '#/components/parameters/LimitTopKeywords'
      responses:
        '200':
          description: OK
          headers:
            X-RateLimit-Limit:
              { $ref: '#/components/headers/X-RateLimit-Limit' }
            X-RateLimit-Remaining:
              { $ref: '#/components/headers/X-RateLimit-Remaining' }
            X-RateLimit-Reset:
              { $ref: '#/components/headers/X-RateLimit-Reset' }
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/KeywordCount' }
              examples:
                ex:
                  value:
                    - { keyword: 'delivery', count: 42 }
                    - { keyword: 'payment', count: 18 }
      x-slos: { response_p95_ms: 500, availability: 99.9 }
      x-grafana: { dashboard: 'TODO: https://grafana.example.com/d/keywords' }
  /analytics/messages-timeline:
    get:
      tags: [analytics]
      operationId: analytics_messages_timeline
      summary: analytics.operations.trends.summary
      description: analytics.operations.trends.description
      parameters:
        - $ref: '#/components/parameters/PeriodParam'
        - $ref: '#/components/parameters/GroupByParam'
      responses:
        '200':
          description: OK
          headers:
            X-RateLimit-Limit:
              { $ref: '#/components/headers/X-RateLimit-Limit' }
            X-RateLimit-Remaining:
              { $ref: '#/components/headers/X-RateLimit-Remaining' }
            X-RateLimit-Reset:
              { $ref: '#/components/headers/X-RateLimit-Reset' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessagesTimelineResponse' }
              examples:
                dayBuckets:
                  value:
                    - { bucket: '2025-09-10', count: 120 }
                    - { bucket: '2025-09-11', count: 95 }
      x-slos: { response_p95_ms: 500, availability: 99.9 }
      x-grafana: { dashboard: 'TODO: https://grafana.example.com/d/timeline' }
...
  /analytics/webhook:
    post:
      tags: [webhooks]
      operationId: analytics_webhook_create_missing
      summary: Webhook للتحليلات (عام)
      description: يستقبل حدثًا لإنشاء MissingResponse من قنوات مختلفة.
      security: [] # @Public()
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateMissingResponseDto' }
            examples:
              whatsappMsg:
                value:
                  merchant: 'merchant_123'
                  channel: 'whatsapp'
                  question: 'هل يتوفر لديكم هذا المنتج؟'
                  botReply: 'عفوًا، لم أفهم السؤال.'
                  sessionId: 'session_xyz'
                  aiAnalysis: 'العميل يستفسر عن توفر منتج.'
                  customerId: 'customer_456'
                  type: 'missing_response'
                  resolved: false
      responses:
        '200':
          description: Acknowledged
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AckWithId' }
              examples:
                ex: { value: { success: true, id: '64f2c1a9e1...' } }
      x-slos: { response_p95_ms: 500, availability: 99.9 }
      x-grafana: { dashboard: 'TODO: https://grafana.example.com/d/webhooks' }
  /analytics/missing-responses:
    get:
      tags: [analytics]
      operationId: analytics_missing_responses_list
      summary: جلب الرسائل المنسية / غير المجاب عنها
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/ResolvedParam'
        - $ref: '#/components/parameters/ChannelParam'
        - $ref: '#/components/parameters/TypeParam'
        - $ref: '#/components/parameters/SearchParam'
        - $ref: '#/components/parameters/FromParam'
        - $ref: '#/components/parameters/ToParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ListResult' }
              examples:
                ex:
                  value:
                    items:
                      - id: '64f2...'
                        channel: 'whatsapp'
                        question: 'ما مقاسات المنتج؟'
                        botReply: '—'
                        resolved: false
                        createdAt: '2025-09-12T10:00:00Z'
                    total: 250
                    page: 1
                    limit: 20
      x-slos: { response_p95_ms: 500, availability: 99.9 }
      x-grafana:
        { dashboard: 'TODO: https://grafana.example.com/d/missing-list' }
  /analytics/missing-responses/{id}/resolve:
    patch:
      tags: [analytics]
      operationId: analytics_missing_responses_resolve_one
      summary: تحديد رسالة كمُعالجة
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ResolveOneResponse' }
      x-slos: { response_p95_ms: 500, availability: 99.9 }
      x-grafana:
        { dashboard: 'TODO: https://grafana.example.com/d/missing-resolve' }
  /analytics/missing-responses/resolve:
    patch:
      tags: [analytics]
      operationId: analytics_missing_responses_resolve_bulk
      summary: تحديد عدة رسائل كمُعالجة
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResolveBulkBody' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ResolveBulkResponse' }
      x-slos: { response_p95_ms: 500, availability: 99.9 }
      x-grafana:
        { dashboard: 'TODO: https://grafana.example.com/d/missing-bulk' }
  /analytics/missing-responses/{id}/add-to-knowledge:
    post:
      tags: [analytics]
      operationId: analytics_missing_add_to_knowledge
      summary: تحويل الرسالة المنسيّة إلى معرفة (FAQ) + وضعها مُعالجة
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AddToKnowledgeDto' }
            examples:
              ar:
                value:
                  {
                    question: 'ماهي سياسة الاستبدال؟',
                    answer: 'الاستبدال متاح خلال 14 يوم مع الفاتورة الأصلية.',
                  }
              en:
                value:
                  {
                    question: "What's the return policy?",
                    answer: 'Returns accepted within 14 days with original receipt.',
                  }
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AddToKnowledgeResult' }
      x-slos: { response_p95_ms: 500, availability: 99.9 }
      x-grafana:
        { dashboard: 'TODO: https://grafana.example.com/d/add-to-knowledge' }
  /analytics/missing-responses/stats:
    get:
      tags: [analytics]
      operationId: analytics_missing_stats
      summary: إحصاءات الرسائل المنسية
      parameters:
        - name: days
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 90, default: 7 }
        - name: notify
          in: query
          required: false
          schema:
            type: string
            enum: ['true', 'false']
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StatsResponse' }
      x-slos: { response_p95_ms: 500, availability: 99.9 }
      x-grafana:
        { dashboard: 'TODO: https://grafana.example.com/d/missing-stats' }
  /analytics/top-products:
    get:
      tags: [analytics]
      operationId: analytics_top_products
      summary: الحصول على أبرز المنتجات
      parameters:
        - $ref: '#/components/parameters/PeriodParam'
        - $ref: '#/components/parameters/LimitTopProducts'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/TopProduct' }
              examples:
                ex:
                  value:
                    - {
                        productId: '64f2...',
                        name: 'Premium Hoodie',
                        count: 17,
                      }
                    - { productId: '64f3...', name: 'Sneakers', count: 9 }
      x-slos: { response_p95_ms: 500, availability: 99.9 }
      x-grafana:
        { dashboard: 'TODO: https://grafana.example.com/d/top-products' }
  /analytics/webhook/kleem:
    post:
      tags: [webhooks]
      operationId: analytics_webhook_kleem_create_missing
      summary: Webhook عام (Kleem)
      security: [] # @Public()
      requestBody:
        required: true
        content:
          application/json:
            schema:
              { $ref: '#/components/schemas/CreateKleemMissingResponseDto' }
            examples:
              kleem:
                value:
                  channel: 'telegram'
                  question: 'متى يصل الطلب؟'
                  botReply: ''
                  sessionId: 'sess_abc'
                  customerId: 'cust_123'
                  aiAnalysis: 'استفسار حالة شحن'
                  resolved: false
                  category: 'shipping'
                  sourceId: 'bot_789'
      responses:
        '200':
          description: Acknowledged
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AckWithId' }
      x-slos: { response_p95_ms: 500, availability: 99.9 }
      x-grafana: { dashboard: 'TODO: https://grafana.example.com/d/webhooks' }
...